// src/modules/invoice/usecase/find-all-invoice/find-all-invoice.dto.ts
export interface FindAllInvoiceOutputDto {
  id: string;
  name: string;
  document: string;
  address: {
    street: string;
    number: number;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
  };
  items: {
    id: string;
    name: string;
    price: number;
  }[];
  total: number;
  createdAt: Date;
}